Splitting the App component
